TAO-8134 - The functionality is extracted to services
