+++ controller/Previewer.php
@@ -20,31 +20,15 @@
 namespace oat\taoQtiTestPreviewer\controller;
 
 use common_Exception;
-use common_Logger;
 use oat\generis\model\OntologyAwareTrait;
 use oat\tao\model\media\sourceStrategy\HttpSource;
 use oat\tao\model\routing\AnnotationReader\security;
 use oat\taoItems\model\media\ItemMediaResolver;
-use oat\taoQtiItem\helpers\QtiFile;
 use oat\taoQtiTestPreviewer\models\ItemPreviewer;
 use oat\taoResultServer\models\classes\ResultServerService;
-use OutOfBoundsException;
-use OutOfRangeException;
-use qtism\common\datatypes\files\FileManagerException;
-use qtism\common\datatypes\files\FileSystemFileManager;
-use qtism\data\storage\StorageException;
-use qtism\data\storage\xml\XmlDocument;
-use qtism\runtime\common\State;
-use qtism\runtime\common\Variable;
 use qtism\runtime\tests\AssessmentItemSession;
-use qtism\runtime\tests\AssessmentItemSessionException;
-use qtism\runtime\tests\SessionManager;
-use RuntimeException;
 use tao_actions_ServiceModule as ServiceModule;
 use taoQtiCommon_helpers_PciStateOutput;
-use taoQtiCommon_helpers_PciVariableFiller;
-use taoQtiCommon_helpers_ResultTransmissionException;
-use taoQtiCommon_helpers_Utils;
 use taoQtiTest_helpers_TestRunnerUtils as TestRunnerUtils;
 use oat\taoItems\model\pack\ItemPack;
 use oat\taoItems\model\pack\Packer;
@@ -211,7 +195,7 @@ class Previewer extends ServiceModule
                 $itemDefinition = $this->getRequestParameter('itemDefinition');
                 $delivery = $this->getResource($this->getRequestParameter('deliveryUri'));
 
-                $itemPreviewer = new ItemPreviewer();
+                $itemPreviewer = $this->getServiceLocator()->get(ItemPreviewer::class);
                 $itemPreviewer->setServiceLocator($this->getServiceLocator());
 
                 $response['content'] = $itemPreviewer->setItemDefinition($itemDefinition)
@@ -278,11 +262,9 @@ class Previewer extends ServiceModule
 
         try {
             $this->validateCsrf();
-
             $itemUri = $this->getPsrRequest()->getQueryParams('itemUri');
-
-            $response = $this->processResponses($itemUri);
-
+            $jsonPayload = $this->getPayload();
+            $response = $this->getItemPreviewer()->processResponses($itemUri, $jsonPayload);
         } catch (\Exception $e) {
             $response = $this->getErrorResponse($e);
             $code = $this->getErrorCode($e);
@@ -292,96 +274,11 @@ class Previewer extends ServiceModule
     }
 
     /**
-     * Item's ResponseProcessing.
-     * 
-     * @param $itemUri
-     * @throws \common_exception_Error
-     * @throws \qtism\common\datatypes\files\FileManagerException
-     * @throws common_Exception
+     * @return ItemPreviewer
      */
-    protected function processResponses($itemUri){
-        if (empty($itemUri)) {
-            throw new common_Exception('missing required itemUri');
-        }
-
-        $item = $this->getResource($itemUri['itemUri']);
-        $qtiXmlDoc = $this->getQtiXmlDoc($item);
-        $itemSession = $this->getItemSession($qtiXmlDoc);
-        $filler = $this->getVariableFilter($qtiXmlDoc);
-        $variables = $this->getQtiSmVariables($filler);
-
-        try {
-            $itemSession->beginAttempt();
-            $itemSession->endAttempt(new State($variables));
-
-            // Return the item session state to the client-side.
-            return [
-                'success' => true,
-                'displayFeedback' => true,
-                'itemSession' => self::buildOutcomeResponse($itemSession)
-            ];
-        }
-        catch(AssessmentItemSessionException $e) {
-            $msg = "An error occurred while processing the responses.";
-            throw new RuntimeExceptionAlias($msg, 0, $e);
-        }
-        catch(taoQtiCommon_helpers_ResultTransmissionException $e) {
-            $msg = "An error occurred while transmitting a result to the target Result Server.";
-            throw new RuntimeExceptionAlias($msg, 0, $e);
-        }
-    }
-
-    protected function buildOutcomeResponse(AssessmentItemSession $itemSession) {
-        $stateOutput = new taoQtiCommon_helpers_PciStateOutput();
-
-        foreach ($itemSession->getOutcomeVariables(false) as $var) {
-            $stateOutput->addVariable($var);
-        }
-
-        $output = $stateOutput->getOutput();
-        return $output;
-    }
-
-    /**
-     * @param XmlDocument $qtiXmlDoc
-     * @return taoQtiCommon_helpers_PciVariableFiller
-     */
-    private function getVariableFilter($qtiXmlDoc)
+    private function getItemPreviewer()
     {
-        return new taoQtiCommon_helpers_PciVariableFiller($qtiXmlDoc->getDocumentComponent());
-    }
-
-    /**
-     * @param XmlDocument $qtiXmlDoc
-     * @return AssessmentItemSession
-     */
-    private function getItemSession($qtiXmlDoc)
-    {
-        $itemSession = new AssessmentItemSession($qtiXmlDoc->getDocumentComponent(), new SessionManager());
-        $itemSession->beginItemSession();
-
-        return $itemSession;
-    }
-
-    /**
-     * @param \core_kernel_classes_Resource $item
-     * @return XmlDocument
-     * @throws common_Exception
-     */
-    private function getQtiXmlDoc($item)
-    {
-        try {
-            $qtiXmlFileContent = QtiFile::getQtiFileContent($item);
-            $qtiXmlDoc = new XmlDocument();
-            $qtiXmlDoc->loadFromString($qtiXmlFileContent);
-        }
-        catch(StorageException $e) {
-            $msg = "An error occurred while loading QTI-XML file at expected location '${qtiXmlFilePath}'.";
-            $this->logError(($e->getPrevious() !== null) ? $e->getPrevious()->getMessage() : $e->getMessage());
-            throw new RuntimeException($msg, 0, $e);
-        }
-
-        return $qtiXmlDoc;
+        return $this->getServiceLocator()->get(ItemPreviewer::class);
     }
 
     /**
@@ -394,44 +291,5 @@ class Previewer extends ServiceModule
         $jsonPayload = json_decode($jsonPayload['itemResponse'], true);
 
         return $jsonPayload;
-    }
-
-    /**
-     * Convert client-side data as QtiSm Runtime Variables
-     * @param taoQtiCommon_helpers_PciVariableFiller $filler
-     * @throws FileManagerException
-     * @return Variable[]
-     */
-    private function getQtiSmVariables($filler)
-    {
-        $variables = array();
-        $jsonPayload = $this->getPayload();
-
-        foreach ($jsonPayload as $id => $response) {
-            try {
-                $var = $filler->fill($id, $response);
-                // Do not take into account QTI Files at preview time.
-                // Simply delete the created file.
-                if (taoQtiCommon_helpers_Utils::isQtiFile($var, false) === true) {
-                    $fileManager = new FileSystemFileManager();
-                    $fileManager->delete($var->getValue());
-                }
-                else {
-                    $variables[] = $var;
-                }
-            }
-            catch (OutOfRangeException $e) {
-                // A variable value could not be converted, ignore it.
-                // Developer's note: QTI Pairs with a single identifier (missing second identifier of the pair) are transmitted as an array of length 1,
-                // this might cause problem. Such "broken" pairs are simply ignored.
-                $this->logDebug("Client-side value for variable '${id}' is ignored due to data malformation.");
-            }
-            catch (OutOfBoundsException $e) {
-                // No such identifier found in item.
-                $this->logDebug("The variable with identifier '${id}' is not declared in the item definition.");
-            }
-        }
-
-        return $variables;
     }
 }
