+++ models/ItemPreviewer.php
@@ -19,20 +19,38 @@
 
 namespace oat\taoQtiTestPreviewer\models;
 
+use common_Exception;
 use common_exception_InconsistentData;
 use common_exception_NotFound;
+use oat\generis\model\OntologyAwareTrait;
+use oat\oatbox\service\ConfigurableService;
 use oat\taoDelivery\model\RuntimeService;
 use oat\taoItems\model\pack\ItemPack;
 use oat\taoItems\model\pack\Packer;
+use oat\taoQtiItem\helpers\QtiFile;
 use oat\taoQtiItem\model\qti\Service;
 use oat\taoQtiItem\model\QtiJsonItemCompiler;
 use oat\taoQtiTest\models\container\QtiTestDeliveryContainer;
-use Zend\ServiceManager\ServiceLocatorAwareInterface;
-use Zend\ServiceManager\ServiceLocatorAwareTrait;
+use OutOfBoundsException;
+use OutOfRangeException;
+use qtism\common\datatypes\files\FileManagerException;
+use qtism\common\datatypes\files\FileSystemFileManager;
+use qtism\data\storage\StorageException;
+use qtism\data\storage\xml\XmlDocument;
+use qtism\runtime\common\State;
+use qtism\runtime\common\Variable;
+use qtism\runtime\tests\AssessmentItemSession;
+use qtism\runtime\tests\AssessmentItemSessionException;
+use qtism\runtime\tests\SessionManager;
+use RuntimeException;
+use taoQtiCommon_helpers_PciStateOutput;
+use taoQtiCommon_helpers_PciVariableFiller;
+use taoQtiCommon_helpers_ResultTransmissionException;
+use taoQtiCommon_helpers_Utils;
 
-class ItemPreviewer implements ServiceLocatorAwareInterface
+class ItemPreviewer extends ConfigurableService
 {
-    use ServiceLocatorAwareTrait;
+    use OntologyAwareTrait;
 
     /**
      * @var string
@@ -192,6 +210,142 @@ class ItemPreviewer implements ServiceLocatorAwareInterface
     {
         return $this->getItemPublicDir()->getPublicAccessUrl() . $this->userLanguage . '/';
     }
+
+    /**
+     * Item's ResponseProcessing.
+     *
+     * @param string $itemUri
+     * @param array $jsonPayload
+     * @return array
+     * @throws FileManagerException
+     * @throws common_Exception
+     */
+    public function processResponses($itemUri, $jsonPayload){
+        if (empty($itemUri)) {
+            throw new common_Exception('missing required itemUri');
+        }
+
+        $item = $this->getResource($itemUri['itemUri']);
+        $qtiXmlDoc = $this->getQtiXmlDoc($item);
+        $itemSession = $this->getItemSession($qtiXmlDoc);
+        $filler = $this->getVariableFilter($qtiXmlDoc);
+        $variables = $this->getQtiSmVariables($filler, $jsonPayload);
+
+        try {
+            $itemSession->beginAttempt();
+            $itemSession->endAttempt(new State($variables));
+
+            // Return the item session state to the client-side.
+            return [
+                'success' => true,
+                'displayFeedback' => true,
+                'itemSession' => self::buildOutcomeResponse($itemSession)
+            ];
+        }
+        catch(AssessmentItemSessionException $e) {
+            $msg = "An error occurred while processing the responses.";
+            throw new RuntimeExceptionAlias($msg, 0, $e);
+        }
+        catch(taoQtiCommon_helpers_ResultTransmissionException $e) {
+            $msg = "An error occurred while transmitting a result to the target Result Server.";
+            throw new RuntimeExceptionAlias($msg, 0, $e);
+        }
+    }
+
+    /**
+     * Convert client-side data as QtiSm Runtime Variables
+     * @param taoQtiCommon_helpers_PciVariableFiller $filler
+     * @throws FileManagerException
+     * @return Variable[]
+     */
+    private function getQtiSmVariables($filler, $jsonPayload)
+    {
+        $variables = array();
+
+        foreach ($jsonPayload as $id => $response) {
+            try {
+                $var = $filler->fill($id, $response);
+                // Do not take into account QTI Files at preview time.
+                // Simply delete the created file.
+                if (taoQtiCommon_helpers_Utils::isQtiFile($var, false) === true) {
+                    $fileManager = new FileSystemFileManager();
+                    $fileManager->delete($var->getValue());
+                }
+                else {
+                    $variables[] = $var;
+                }
+            }
+            catch (OutOfRangeException $e) {
+                // A variable value could not be converted, ignore it.
+                // Developer's note: QTI Pairs with a single identifier (missing second identifier of the pair) are transmitted as an array of length 1,
+                // this might cause problem. Such "broken" pairs are simply ignored.
+                $this->logDebug("Client-side value for variable '${id}' is ignored due to data malformation.");
+            }
+            catch (OutOfBoundsException $e) {
+                // No such identifier found in item.
+                $this->logDebug("The variable with identifier '${id}' is not declared in the item definition.");
+            }
+        }
+
+        return $variables;
+    }
+
+
+    /**
+     * @param XmlDocument $qtiXmlDoc
+     * @return taoQtiCommon_helpers_PciVariableFiller
+     */
+    private function getVariableFilter($qtiXmlDoc)
+    {
+        return new taoQtiCommon_helpers_PciVariableFiller($qtiXmlDoc->getDocumentComponent());
+    }
+
+    /**
+     * @param XmlDocument $qtiXmlDoc
+     * @return AssessmentItemSession
+     */
+    private function getItemSession($qtiXmlDoc)
+    {
+        $itemSession = new AssessmentItemSession($qtiXmlDoc->getDocumentComponent(), new SessionManager());
+        $itemSession->beginItemSession();
+
+        return $itemSession;
+    }
+
+    /**
+     * @param \core_kernel_classes_Resource $item
+     * @return XmlDocument
+     * @throws common_Exception
+     */
+    private function getQtiXmlDoc($item)
+    {
+        try {
+            $qtiXmlFileContent = QtiFile::getQtiFileContent($item);
+            $qtiXmlDoc = new XmlDocument();
+            $qtiXmlDoc->loadFromString($qtiXmlFileContent);
+        }
+        catch(StorageException $e) {
+            $msg = "An error occurred while loading QTI-XML file at expected location '${qtiXmlFilePath}'.";
+            $this->logError(($e->getPrevious() !== null) ? $e->getPrevious()->getMessage() : $e->getMessage());
+            throw new RuntimeException($msg, 0, $e);
+        }
+
+        return $qtiXmlDoc;
+    }
+
+    protected function buildOutcomeResponse(AssessmentItemSession $itemSession) {
+        $stateOutput = new taoQtiCommon_helpers_PciStateOutput();
+
+        foreach ($itemSession->getOutcomeVariables(false) as $var) {
+            $stateOutput->addVariable($var);
+        }
+
+        $output = $stateOutput->getOutput();
+        return $output;
+    }
+
+
+
 
     /**
      * @return \tao_models_classes_service_StorageDirectory
